requirements.txt
Django==4.2.5
django-extensions==3.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
pytest==7.4.2
pytest-django==4.5.2
drf-yasg==1.21.7
